refactor(ui): Phase 1 — extract shared SidebarChat modules
Split icons, path helpers, and tool primitives out of the 5k-line
SidebarChat.tsx so consumers no longer depend on the monolith, with
pathDisplay unit tests and re-exports for compatibility.
