+++ src/vs/workbench/contrib/cortexide/browser/react/src/markdown/ApplyBlockHoverButtons.tsx
@@ -10,7 +10,8 @@ import { isFeatureNameDisabled } from '../../../../common/cortexideSettingsTypes
 import { URI } from '../../../../../../../base/common/uri.js'
 import { FileSymlink, LucideIcon, RotateCw, Terminal } from 'lucide-react'
 import { Check, X, Square, Copy, Play, } from 'lucide-react'
-import { getBasename, ListableToolItem, voidOpenFileFn, ToolChildrenWrapper } from '../sidebar-tsx/SidebarChat.js'
+import { getBasename, voidOpenFileFn } from '../sidebar-tsx/shared/pathUtils.js'
+import { ListableToolItem, ToolChildrenWrapper } from '../sidebar-tsx/tools/ToolPrimitives.js'
 import { PlacesType, VariantType } from 'react-tooltip'
 
 enum CopyButtonText {
